backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
sentencepiece
python-dotenv
pydantic
requests
python-multipart
pytesseract
Pillow